Reorganizing enum files
