+++ backend/src/config/env.ts
@@ -3,7 +3,7 @@ import dotenv from 'dotenv';
 import path from 'path';
 import { fileURLToPath } from 'url';
 
-import { ErrorCode } from '../types/ErrorCode.js';
+import { ErrorCode } from '../enums/ErrorCode.js';
 import { AppError } from '../utils/errors.js';
 
 /**
