+++ backend/src/enums/Environment.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 export enum ENVIRONMENT {
   DEV = 'dev',
   INT = 'int',
