+++ backend/src/enums/ErrorCode.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 export enum ErrorCode {
   INTERNAL_ERROR = 'INTERNAL_ERROR',
   BAD_REQUEST = 'BAD_REQUEST',
