+++ backend/src/enums/HttpMethod.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 export enum HttpMethod {
   GET = 'GET',
   POST = 'POST',
