+++ backend/src/enums/HttpStatusCode.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 export enum HttpStatusCode {
   OK = 200,
   BAD_REQUEST = 400,
